backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
langchain
langchain-groq
langgraph
pydantic
groq

